Added Hand sprite changing
HoldingPoint now accesses the Equipment manager and uses inputs to render a held item on screen. It creates a Polygon collider for the item when necessary. Not sure how this will work with animations. Also added slappy arm spritesheet. May want to add wind-up and wind-down animations for melee weapons.