More work on shield mechanics.
Changed Holding point to allow the shield key to bring out the shield as
expected. The shield cannot be brought out while attacking, but weapons
can be swapped while shielding.

Created a new interface into the EquipmentManager for other classes to
get the current equipment without directly accessing the array.