Enemy GameObjects now follow DOTS paths to completion
Enemy scripts can now request a path from the PathfindingGrid.
The pathfinding system will then populate their attached entity with the
path, and the enemy will follow the path to completion.

The enemy can be interrupted with another path, but it doesn't look
great.

This interruption is also manual. The next step is to automatically
update the path.