MovementController rework and Pathfinding bugfix
Reworked MovementController so that only one call to Travel() is
necessary to move an entity across the map.

Fixed a bug in Pathfinding where turns on the last square in a path were
bypassing the InSight check, causing paths through walls. Also commented
the shit out of Waypointification because it's incredibly complicated.