Starting DOTS removal
DOTS has proved to be very complex to integrate with OOP. I've decided
to remove it for now. It can be added back whenever.

This creates a central VoraUtils namespace for all of these functions.
It includes an Astar class. This commit prepares for the complete
removal of DOTS and the substitution with steering behaviors.