Added JPS (no plus) pathfinding
JPS is much simpler as it is just a successor filter. Added this filter
to PathfindingGrid. Left old filter in, just unused. Fixed a few bugs
in the waypointifying, and made the turn finder independant of JPS or
normal A-star.

Fixed GetWorldPosition to factor in the cell size.

Added operators to Coordinate.