Added speeding up/slowing down on movementcontroller and knockback for damaging
New ChangeMaxSpeedOverTime function slowly changes the max speed of an entity, leading to much more natural stopping/slowing down.

New ApplyKnockback function to player moves him back a bit. It may be worth creating an interface, as this will be something that happens in a variety of circumstances.

Strange bug: every so often, the knockback is about ten times stronger than it should be  and the player teleports.